+++ back/app/helpers/searches_helper.rb
@@ -1,0 +1,2 @@
+module SearchesHelper
+end
