+++ back/app/services/movie_fetcher.rb
@@ -6,11 +6,11 @@ class MovieFetcher
 
   def self.movie_data_detail(movie_id)
     movie_data = fetch_movie_data(movie_id, "ja")
-
-    if movie_data[:overview].nil? || movie_data[:overview].empty?
+    Rails.logger.debug("ttttttttttttttttttt#{movie_data}")
+    if movie_data["overview"].nil? || movie_data["overview"].empty?
       movie_data = fetch_movie_data(movie_id, "en")
-      # translated_overview = translate_text(movie_data[:overview], "ja")
-      # movie_data[:overview] = "#{translated_overview}(＊英文を翻訳した内容なので表現に誤りがある場合があります)"
+      translated_overview = translate_text(movie_data["overview"], "ja")
+      movie_data["overview"] = "#{translated_overview}(＊英文を翻訳した内容なので表現に誤りがある場合があります)"
     end
 
     movie_data
@@ -33,7 +33,7 @@ class MovieFetcher
   def self.translate_text(text, change_language)
     translate = Google::Cloud::Translate::V2.new(
       project_id: ENV['GOOGLE_PROJECT_ID'],
-      credentials:  Rails.application.credentials.api_key[:google],
+      credentials: JSON.parse(File.read(ENV['GOOGLE_API'])),
     )
 
     translation = translate.translate text, to: change_language
