+++ back/app/views/searches/search.html.erb
@@ -1,6 +1,6 @@
 <h1>検索</h1>
 
-<%= form_with url: search_movies_path, method: :get, local: true do %>
+<%= form_with url: search_path, method: :get, local: true do %>
   <%= label_tag :query, "映画検索" %>
   <%= text_field_tag :query, params[:query] %>
   <%= submit_tag "検索" %>
