+++ back/app/views/shared/_header.html.erb
@@ -8,7 +8,7 @@
       <ul class="navbar-nav">
         <li class="nav-item">
           <%= link_to "ランダム", movies_path %>
-          <%= link_to "検索", search_movies_path %>
+          <%= link_to "検索", search_path %>
           <%= link_to "マイページ", profile_path %>
           <%= link_to "特集", collections_path %>
           <%= link_to 'ログアウト', destroy_user_session_path, class: 'nav-link', method: :delete %>
